Reorganize MPI sources
