+++ src/CMakeLists.txt
@@ -17,10 +17,6 @@ add_subdirectory(reader)
 add_subdirectory(services)
 add_subdirectory(caliper)
 
-if (CALIPER_HAVE_MPI)
-  add_subdirectory(mpi)
-endif()
-
 set(CALIPER_COMMON_OBJS
   $<TARGET_OBJECTS:caliper-common>
   $<TARGET_OBJECTS:caliper-interface>
@@ -30,37 +26,8 @@ set(CALIPER_COMMON_OBJS
   ${CALIPER_SERVICES_LIBS}
   ${CALIPER_EXTERNAL_SOURCES})
 
-list(APPEND CALIPER_SERIAL_OBJS
-  ${CALIPER_COMMON_OBJS})
 list(APPEND CALIPER_ALL_OBJS
   ${CALIPER_COMMON_OBJS})
-
-# Make (non-mpi) libcaliper-serial
-
-add_library(caliper-serial ${CALIPER_SERIAL_OBJS})
-
-set_target_properties(caliper-serial PROPERTIES SOVERSION ${CALIPER_MAJOR_VERSION})
-set_target_properties(caliper-serial PROPERTIES VERSION ${CALIPER_VERSION})
-
-target_compile_features(caliper-serial PUBLIC cxx_std_11)
-
-target_include_directories(
-  caliper-serial
-  INTERFACE
-  "$<INSTALL_INTERFACE:include>")
-
-list(APPEND CALIPER_EXTERNAL_LIBS
-  Threads::Threads)
-
-foreach (_extlib ${CALIPER_EXTERNAL_LIBS})
-  target_link_libraries(caliper-serial PRIVATE ${_extlib})
-endforeach()
-
-if (WIN32)
-  target_link_libraries(caliper-serial PUBLIC ws2_32)
-endif()
-
-# Make the combined libcaliper
 
 add_library(caliper ${CALIPER_ALL_OBJS})
 
@@ -88,11 +55,6 @@ endif()
 if (WITH_TOOLS)
   add_subdirectory(tools)
 endif()
-
-install(TARGETS caliper-serial
-  EXPORT caliper-targets
-  ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
-  LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR})
 
 install(TARGETS caliper
   EXPORT caliper-targets
