+++ src/caliper/CMakeLists.txt
@@ -16,6 +16,22 @@ set(CALIPER_RUNTIME_SOURCES
   cali_datatracker.cpp
   config_sanity_check.cpp)
 
+if (CALIPER_HAVE_MPI)
+  list(APPEND CALIPER_RUNTIME_SOURCES
+    aggregate_over_mpi.cpp
+    collective_flush.cpp
+    CollectiveOutputChannel.cpp
+    OutputCommMpi.cpp
+    MpiChannelManager.cpp
+    mpi_flush.cpp
+    machine_mpi.cpp
+    setup_mpi.cpp)
+else()
+  list(APPEND CALIPER_RUNTIME_SOURCES
+    machine_serial.cpp
+    setup_serial.cpp)
+endif()
+
 add_subdirectory(controllers)
 
 add_library(caliper-runtime OBJECT ${CALIPER_RUNTIME_SOURCES})
@@ -23,23 +39,9 @@ add_library(caliper-runtime OBJECT ${CALIPER_RUNTIME_SOURCES})
 target_compile_options(caliper-runtime PRIVATE ${Wall_flag})
 target_compile_features(caliper-runtime PUBLIC cxx_std_11)
 
-list(APPEND CALIPER_SERIAL_OBJS
-  caliper/machine_serial.cpp
-  caliper/setup_serial.cpp)
-
 if (CALIPER_HAVE_MPI)
   target_include_directories(caliper-runtime PRIVATE ${MPI_CXX_INCLUDE_PATH})
-  list(APPEND CALIPER_ALL_OBJS
-    caliper/machine_mpi.cpp
-    caliper/setup_mpi.cpp)
-else()
-  list(APPEND CALIPER_ALL_OBJS
-    caliper/machine_serial.cpp
-    caliper/setup_serial.cpp)
 endif()
-
-set(CALIPER_SERIAL_OBJS ${CALIPER_SERIAL_OBJS} PARENT_SCOPE)
-set(CALIPER_ALL_OBJS    ${CALIPER_ALL_OBJS}    PARENT_SCOPE)
 
 if (BUILD_TESTING)
   add_subdirectory(test)
