+++ src/caliper/CollectiveOutputChannel.cpp
