+++ src/caliper/MpiChannelManager.cpp
