+++ src/caliper/machine_mpi.cpp
@@ -7,9 +7,58 @@
 
 #include "caliper/common/Log.h"
 
-#ifdef CALIPER_HAVE_MPI
-#include "../mpi/mpi_machine.h"
-#endif
+#include <mpi.h>
+
+#include <unistd.h>
+
+#include <cstdint>
+#include <string>
+#include <vector>
+
+namespace
+{
+
+int get_rank_for_hash(uint64_t hash, MPI_Comm comm)
+{
+    int myrank = 0;
+    int ccsize = 0;
+
+    PMPI_Comm_rank(comm, &myrank);
+    PMPI_Comm_size(comm, &ccsize);
+
+    std::vector<uint64_t> res(ccsize, 0);
+
+    int ret = PMPI_Allgather(&hash,      1, MPI_UINT64_T,
+                             res.data(), 1, MPI_UINT64_T, comm);
+
+    if (ret != MPI_SUCCESS)
+        return -1;
+
+    int hrank = 0;
+
+    for (int i = 0; i < myrank; ++i)
+        if (hash == res[i])
+            ++hrank;
+
+    return hrank;
+}
+
+int get_rank_for_node()
+{
+    constexpr std::size_t max_len = 1024;
+    char hostname[max_len];
+    if (gethostname(hostname, max_len) < 0)
+        return -1;
+
+    int initialized = 0;
+    PMPI_Initialized(&initialized);
+    if (!initialized)
+        return 0;
+
+    return ::get_rank_for_hash(std::hash<std::string>{}(std::string(hostname)), MPI_COMM_WORLD);
+}
+
+} // namespace [anonymous]
 
 namespace cali
 {
@@ -23,11 +72,7 @@ int get_rank_for(MachineLevel level)
     case MachineLevel::Process:
         return 0;
     case MachineLevel::Node:
-#ifdef CALIPER_HAVE_MPI
-        return mpi::get_rank_for_node();
-#else
-        return 0;
-#endif
+        return ::get_rank_for_node();
     default:
         Log(0).stream() << "machine::get_rank_for(MachineLevel): level "
                         << level << " is not supported"
@@ -37,6 +82,6 @@ int get_rank_for(MachineLevel level)
     return -1;
 }
 
-}
+} // namespace machine
 
 }
