--- src/mpi/README.md
@@ -1,5 +1,0 @@
-MPI-specific services
-========================================
-
-This directory contains the MPI-specific parts of the Caliper runtime system,
-as well as MPI-based services.
