--- src/mpi/services/mpireport/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_mpi_service_sources(MpiReport.cpp)
