+++ src/services/CMakeLists.txt
@@ -115,6 +115,14 @@ endif()
 if (CALIPER_HAVE_CRAYPAT)
   add_subdirectory(craypat)
 endif()
+if (CALIPER_HAVE_MPI)
+  include_directories(${MPI_CXX_INCLUDE_PATH})
+  add_subdirectory(mpireport)
+  add_subdirectory(mpiwrap)
+  if (CALIPER_HAVE_TAU)
+    add_subdirectory(tau)
+  endif()
+endif()
 
 if (CALIPER_BUILD_TESTING)
   add_subdirectory(templates)
