+++ src/services/mpit/mpitdesign/design.tex
