+++ src/tools/cali-query/CMakeLists.txt
@@ -17,7 +17,7 @@ add_executable(cali-query
   $<TARGET_OBJECTS:query-common>
   ${CALIPER_QUERY_SOURCES})
 
-target_link_libraries(cali-query caliper-tools-util caliper-serial)
+target_link_libraries(cali-query caliper-tools-util caliper)
 target_link_libraries(cali-query Threads::Threads)
 
 install(TARGETS cali-query DESTINATION ${CMAKE_INSTALL_BINDIR})
