+++ src/tools/cali-stat/CMakeLists.txt
@@ -3,6 +3,6 @@ set(CALIPER_STAT_SOURCES
 
 add_executable(cali-stat ${CALIPER_STAT_SOURCES})
 
-target_link_libraries(cali-stat caliper-tools-util caliper-serial)
+target_link_libraries(cali-stat caliper-tools-util caliper)
 
 install(TARGETS cali-stat DESTINATION ${CMAKE_INSTALL_BINDIR})
